python/pyproject.toml
[project]
name = "entertainment-graph"
version = "0.1.0"
description = "Graph RAG system for entertainment discovery"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",

    # LLM & Embeddings
    "openai>=1.0.0",

    # Vector Store (baseline)
    "chromadb>=0.4.0",

    # Graph Systems
    "graphiti-core>=0.3.0",
    "neo4j>=5.0.0",

    # Memory Systems
    "openmemory-py>=0.1.0",

    # HTTP client
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
